Calculate follow-up time per case until end of follow-up depending on pat_status - tidyverse version

Calculate follow-up time per case until end of follow-up depending on pat_status - tidyverse version

calc_futime(
  wide_df,
  futime_var_new = "p_futimeyrs",
  fu_end,
  dattype = "zfkd",
  check = TRUE,
  time_unit = "years",
  status_var = "p_status",
  lifedat_var = NULL,
  fcdat_var = NULL,
  spcdat_var = NULL
)

Arguments

wide_df
dataframe in wide format

futime_var_new
Name of the newly calculated variable for follow-up time. Default is p_futimeyrs.

fu_end
end of follow-up in time format YYYY-MM-DD.

dattype
Type of cancer registry data. Can be "seer" or "zfkd". Default is "zfkd".

check
Check newly calculated variable p_status by printing frequency table. Default is TRUE.

time_unit
Unit of follow-up time (can be "days", "weeks", "months", "years"). Default is "years".

status_var
Name of the patient status variable that was previously created. Default is p_status.

lifedat_var
Name of variable containing Date of Death. Will override dattype preset.

fcdat_var
Name of variable containing Date of Primary Cancer diagnosis. Will override dattype preset.

spcdat_var
Name of variable containing Date of SPC diagnosis Will override dattype preset.

Value

wide_df