Calculate age-, sex-, cohort-, region-specific incidence rates from a cohort

Calculate age-, sex-, cohort-, region-specific incidence rates from a cohort

calc_refrates(
  df,
  dattype = "zfkd",
  count_var,
  refpop_df,
  calc_totals = TRUE,
  fill_sites = "no",
  region_var = NULL,
  age_var = NULL,
  sex_var = NULL,
  year_var = NULL,
  race_var = NULL,
  site_var = NULL
)

Arguments

df

dataframe in long format

dattype

can be "zfkd" or "seer" or empty. Will set default variable names from dataset.

count_var

variable to be counted as observed case. Should be 1 for case to be counted.

refpop_df

df where reference population data is defined. Only required if option futime = "refpop" is chosen. It is assumed that refpop_df has the columns "region" for region, "sex" for biological sex, "age" for age-groups (can be single ages or 5-year brackets), "year" for time period (can be single year or 5-year brackets), "population_pyar" for person-years at risk in the respective age/sex/year cohort. refpop_df must use the same category coding of age, sex, region, year and site as age_var, sex_var, region_var, year_var and site_var.

calc_totals

option to calculate totals for all age-groups, all sexes, all years, all races

fill_sites

option to fill missing sites in observed with incidence rate of 0. Needs to define the coding system used. Can be either "no" for not filling missing sites. "icd2d" for ICD-10 2 digit, "icd3d" for ICD-10 3digit, "sitewho" for Site WHO

region_var

variable in df that contains information on region where case was incident. Default is set if dattype is given.

age_var

variable in df that contains information on age-group. Default is set if dattype is given.

sex_var

variable in df that contains information on sex. Default is set if dattype is given.

year_var

variable in df that contains information on year or year-period when case was incident. Default is set if dattype is given.

race_var

optional argument for dattype="seer", if rates should be calculated by race. If you want to use this option, provide variable name of df that contains race information.

site_var

variable in df that contains information on ICD code of case diagnosis. Cases are usually the second cancers. Default is set if dattype is given.

Value

df

Examples

#load sample data
data("us_second_cancer")
data("population_us")

us_second_cancer %>%
  #create variable to indicate to be counted as case
  dplyr::mutate(is_case = 1) %>%
  #calculate refrates - warning: these are not realistic numbers, just showing functionality
  calc_refrates(dattype = "seer", , count_var = "is_case", refpop_df = population_us,
               region_var = "registry", age_var = "fc_agegroup", sex_var = "sex", 
               site_var = "t_site_icd")
#> 
#>  [INFO Ref population missing]For the following age groups, sex, regions, years, t_sites no population can be found: 
#> # A tidytable: 144 × 4
#>    age     sex    region                                   year       
#>    <chr>   <chr>  <chr>                                    <chr>      
#>  1 65 - 69 Male   SEER Reg 20 - Detroit (Metropolitan)     2015 - 2019
#>  2 55 - 59 Female SEER Reg 02 - Connecticut                2015 - 2019
#>  3 65 - 69 Female SEER Reg 20 - Detroit (Metropolitan)     2015 - 2019
#>  4 50 - 54 Male   SEER Reg 21 - Hawaii                     2015 - 2019
#>  5 45 - 49 Male   SEER Reg 02 - Connecticut                2015 - 2019
#>  6 50 - 54 Male   SEER Reg 01 - San Francisco-Oakland SMSA 2015 - 2019
#>  7 65 - 69 Male   SEER Reg 21 - Hawaii                     2015 - 2019
#>  8 75 - 79 Female SEER Reg 01 - San Francisco-Oakland SMSA 2015 - 2019
#>  9 65 - 69 Female SEER Reg 02 - Connecticut                2015 - 2019
#> 10 35 - 39 Male   SEER Reg 01 - San Francisco-Oakland SMSA 2015 - 2019
#> # … with 134 more rows 
#> Check attribute `problems_missing_ref_strata` of results to see what strata are affected. 
#> Solution could be to add these strata to refpop_df. 
#> # A tidytable: 6,181 × 9
#>    t_site region           year    sex   age   incidence_cases incidence_crude_…
#>    <chr>  <fct>            <chr>   <fct> <chr>           <dbl>             <dbl>
#>  1 C14    SEER Reg 01 - S… 1990 -… Fema… 00 -…               1             0.145
#>  2 C14    SEER Reg 01 - S… 1990 -… Fema… 25 -…               1             0.118
#>  3 C14    SEER Reg 01 - S… 1990 -… Fema… 35 -…               2             0.231
#>  4 C14    SEER Reg 01 - S… 1990 -… Fema… 45 -…               3             0.486
#>  5 C14    SEER Reg 01 - S… 1990 -… Fema… 50 -…               1             0.208
#>  6 C14    SEER Reg 01 - S… 1990 -… Fema… 60 -…               2             0.518
#>  7 C14    SEER Reg 01 - S… 1990 -… Fema… 70 -…               1             0.303
#>  8 C14    SEER Reg 01 - S… 1990 -… Fema… 75 -…               1             0.381
#>  9 C14    SEER Reg 01 - S… 1990 -… Male  00 -…               1             0.138
#> 10 C14    SEER Reg 01 - S… 1990 -… Male  25 -…               1             0.113
#> # … with 6,171 more rows, and 2 more variables: population_pyar <dbl>,
#> #   population_n_per_year <dbl>