Calculate crude incidence rates and cross-tabulate results by break variables; cumulative FU-times as are used as xbreak_var

Calculate crude incidence rates and cross-tabulate results by break variables; cumulative FU-times as are used as xbreak_var

ir_crosstab_byfutime(
  df,
  dattype = "zfkd",
  count_var,
  futime_breaks = c(0, 0.5, 1, 5, 10, Inf),
  ybreak_vars,
  collapse_ci = FALSE,
  add_total = "no",
  futime_var = NULL,
  alpha = 0.05
)

Arguments

df
dataframe in wide format

dattype
can be "zfkd" or "seer" or empty. Will set default variable names from dataset.

count_var
variable to be counted as observed case. Should be 1 for case to be counted.

futime_breaks
vector that indicates split points for follow-up time groups (in years) that will be used as xbreak_var. Default is c(0, .5, 1, 5, 10, Inf) that will result in 5 groups (up to 6 months, 6-12 months, 1-5 years, 5-10 years, 10+ years).

ybreak_vars
variables from df by which rates should be stratified in rows of result df. Multiple variables will result in appended rows in result df. y_break_vars is required.

collapse_ci
If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.

add_total
option to add a row of totals. Can be either "no" for not adding such a row or "top" or "bottom" for adding it at the first or last row. Default is "no".

futime_var
variable in df that contains follow-up time per person (in years). Default is set if dattype is given.

alpha
significance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.

Value

df

Examples

#load sample data
data("us_second_cancer")

#prep step - make wide data as this is the required format
usdata_wide <- us_second_cancer %>%
                    #only use sample
                    dplyr::filter(as.numeric(fake_id) < 300000) %>%
                    msSPChelpR::reshape_wide_tidyr(case_id_var = "fake_id", 
                    time_id_var = "SEQ_NUM", timevar_max = 2)

#> Long dataset had too many cases per patient. Wide dataset is limited to  2  cases per id as defined in timevar_max option.

                    
#prep step - calculate p_spc variable
usdata_wide <- usdata_wide %>%
                 dplyr::mutate(p_spc = dplyr::case_when(is.na(t_site_icd.2)   ~ "No SPC",
                                                       !is.na(t_site_icd.2)   ~ "SPC developed",
                                                       TRUE ~ NA_character_)) %>%
                 dplyr::mutate(count_spc = dplyr::case_when(is.na(t_site_icd.2)   ~ 1,
                                                              TRUE ~ 0))
                                                              
#prep step - create patient status variable
usdata_wide <- usdata_wide %>%
                  msSPChelpR::pat_status(., fu_end = "2017-12-31", dattype = "seer",
                                         status_var = "p_status", life_var = "p_alive.1",
                                         birthdat_var = "datebirth.1", lifedat_var = "datedeath.1")

#> # A tibble: 10 x 3
#>    p_alive.1 p_status     n
#>    <chr>        <dbl> <int>
#>  1 Alive            1  3477
#>  2 Alive            2  3892
#>  3 Alive           97     1
#>  4 Alive           98   529
#>  5 Dead             1   537
#>  6 Dead             2   481
#>  7 Dead             3  4052
#>  8 Dead             4  1975
#>  9 Dead            97     1
#> 10 Dead            98    52
#> # A tibble: 6 x 2
#>   p_status     n
#>      <dbl> <int>
#> 1        1  4014
#> 2        2  4373
#> 3        3  4052
#> 4        4  1975
#> 5       97     2
#> 6       98   581

 
#now we can run the function
usdata_wide <- usdata_wide %>%
                 msSPChelpR::calc_futime(., 
                        futime_var_new = "p_futimeyrs", 
                        fu_end = "2017-12-31",
                        dattype = "seer", 
                        time_unit = "years",
                        status_var = "p_status",
                        lifedat_var = "datedeath.1", 
                        fcdat_var = "t_datediag.1", 
                        spcdat_var = "t_datediag.2")

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> # A tibble: 6 x 5
#>   p_status mean_futime min_futime max_futime median_futime
#>      <dbl>       <dbl>      <dbl>      <dbl>         <dbl>
#> 1        1        9.08     0.0438       27.0          7.96
#> 2        2        8.97     0            26.8          7.84
#> 3        3        8.43     0            25.5          7.46
#> 4        4        6.34     0            25.5          5.17
#> 5       97      NaN      Inf          -Inf           NA   
#> 6       98      NaN      Inf          -Inf           NA   

                    
#for example, you can calculate incidence and summarize by sex and registry
msSPChelpR::ir_crosstab_byfutime(usdata_wide,
      dattype = "seer",
      count_var = "count_spc",
      futime_breaks = c(0, .5, 1, 5, 10, Inf),
      ybreak_vars = c("sex.1", "registry.1"),
      collapse_ci = FALSE,
      add_total = "no",
      futime_var = "p_futimeyrs",
      alpha = 0.05)

#> Warning: The variable for follow-up time has: 600 missings. These will be omitted when creating the crosstabs.

#>    yvar_name                               yvar_label to 6 months_n_base
#> 1      sex.1                                   Female               7240
#> 2      sex.1                                     Male               7157
#> 3 registry.1 SEER Reg 01 - San Francisco-Oakland SMSA               4792
#> 4 registry.1                SEER Reg 02 - Connecticut               3184
#> 5 registry.1     SEER Reg 20 - Detroit (Metropolitan)               4813
#> 6 registry.1                     SEER Reg 21 - Hawaii               1608
#>   to 6 months_observed to 6 months_pyar to 6 months_abs_ir
#> 1                  201             3513            5722.34
#> 2                  250             3457            7232.69
#> 3                  137             2328            5883.70
#> 4                  121             1533            7894.92
#> 5                  142             2333            6087.62
#> 6                   51              775            6577.49
#>   to 6 months_abs_ir_lci to 6 months_abs_ir_uci 6-12 months_n_base
#> 1                4958.56                6570.48               6884
#> 2                6363.84                8187.05               6753
#> 3                4939.75                6955.49               4573
#> 4                6550.99                9433.42               2985
#> 5                5127.55                7175.24               4571
#> 6                4897.37                8648.19               1508
#>   6-12 months_observed 6-12 months_pyar 6-12 months_abs_ir
#> 1                  127             3370            3768.50
#> 2                  158             3305            4781.08
#> 3                   85             2242            3790.97
#> 4                   76             1456            5220.35
#> 5                   98             2238            4378.60
#> 6                   26              739            3520.37
#>   6-12 months_abs_ir_lci 6-12 months_abs_ir_uci 1-5 years_n_base
#> 1                3141.63                4483.81             6590
#> 2                4064.64                5587.42             6458
#> 3                3028.09                4687.59             4394
#> 4                4113.04                6534.04             2831
#> 5                3554.76                5336.11             4376
#> 6                2299.63                5158.16             1447
#>   1-5 years_observed 1-5 years_pyar 1-5 years_abs_ir 1-5 years_abs_ir_lci
#> 1                937          22336          4194.95              3930.61
#> 2               1001          21763          4599.58              4319.02
#> 3                638          14953          4266.81              3942.11
#> 4                419           9568          4379.08              3969.77
#> 5                673          14752          4562.10              4223.89
#> 6                208           4826          4309.58              3743.78
#>   1-5 years_abs_ir_uci 5-10 years_n_base 5-10 years_observed 5-10 years_pyar
#> 1              4472.38              4644                1066           17978
#> 2              4893.58              4509                1025           17460
#> 3              4611.13              3126                 723           12153
#> 4              4819.14              1974                 430            7683
#> 5              4920.18              3050                 702           11751
#> 6              4936.75              1003                 236            3851
#>   5-10 years_abs_ir 5-10 years_abs_ir_lci 5-10 years_abs_ir_uci
#> 1           5929.63               5578.97               6296.56
#> 2           5870.43               5516.50               6241.10
#> 3           5949.37               5523.55               6399.30
#> 4           5596.45               5079.92               6151.26
#> 5           5974.04               5540.23               6432.79
#> 6           6128.17               5371.20               6961.94
#>   10+ years_n_base 10+ years_observed 10+ years_pyar 10+ years_abs_ir
#> 1             2613               1350          14539          9285.63
#> 2             2576               1344          14481          9281.14
#> 3             1771                907           9819          9237.63
#> 4             1129                603           6410          9407.88
#> 5             1736                908           9701          9360.34
#> 6              553                276           3091          8929.08
#>   10+ years_abs_ir_lci 10+ years_abs_ir_uci
#> 1              8796.85              9794.50
#> 2              8791.52              9790.93
#> 3              8646.15              9858.92
#> 4              8671.86             10189.66
#> 5              8761.33              9989.52
#> 6              7906.63             10047.06