Calculate patient status at specific end of follow-up - data.table version

Calculate patient status at specific end of follow-up - data.table version

pat_status_dt(
  wide_dt,
  fu_end = NULL,
  dattype = "zfkd",
  status_var = "p_status",
  life_var = NULL,
  spc_var = NULL,
  birthdat_var = NULL,
  lifedat_var = NULL,
  lifedatmin_var = NULL,
  fcdat_var = NULL,
  spcdat_var = NULL,
  life_stat_alive = NULL,
  life_stat_dead = NULL,
  spc_stat_yes = NULL,
  spc_stat_no = NULL,
  lifedat_fu_end = NULL,
  use_lifedatmin = FALSE,
  check = TRUE,
  as_labelled_factor = FALSE
)

Arguments

wide_dt
wide data.table (data.frame can also be provided, but will be converted to data.table)

fu_end
end of follow-up in time format YYYY-MM-DD.

dattype
Type of cancer registry data. Can be "seer" or "zfkd". Default is "zfkd".

status_var
Name of the newly calculated variable for patient status. Default is p_status.

life_var
Name of variable containing life status. Will override dattype preset.

spc_var
Name of variable containing SPC status. Will override dattype preset.

birthdat_var
Name of variable containing Date of Birth. Will override dattype preset.

lifedat_var
Name of variable containing Date of Death. Will override dattype preset.

lifedatmin_var
Name of variable containing the minimum Date of Death when true DoD is missing. Will override dattype preset. Will only be used if use_lifedatmin = TRUE.

fcdat_var
Name of variable containing Date of Primary Cancer diagnosis. Will override dattype preset.

spcdat_var
Name of variable containing Date of SPC diagnosis Will override dattype preset.

life_stat_alive
Value for alive status in life_var. Will override dattype preset.

life_stat_dead
Value for dead status in life_var. Will override dattype preset.

spc_stat_yes
Value for SPC occured in spc_var. Will override dattype preset.

spc_stat_no
Value for no SPC in spc_var. Will override dattype preset.

lifedat_fu_end
Date of last FU of alive status in registry data. Will override dattype preset (2017-03-31 for zfkd; 2018-12-31 for seer).

use_lifedatmin
If TRUE, option to use Date of Death from lifedatmin_var when DOD is missing. Default is FALSE.

check
Check newly calculated variable p_status. Default is TRUE.

as_labelled_factor
If TRUE, output status_var as labelled factor variable. Default is FALSE.

Value

wide_dt