Summarize detailed SIR results

Summarize detailed SIR results

summarize_sir_results(
  sir_df,
  summarize_groups,
  summarize_icdcat,
  output = "long",
  output_information = "full",
  add_total_row = "no",
  add_total_fu = "no",
  collapse_ci = FALSE,
  shorten_total_cols = FALSE,
  fubreak_var_name = "fu_time",
  ybreak_var_name = "yvar_name",
  xbreak_var_name = "none",
  alpha = 0.05
)

Arguments

sir_df
dataframe with stratified sir results created using the sir or sir_byfutime functions

summarize_groups
option to define summarizing stratified groups. Default is "none". If you want to define variables that should be summarized into one group, you can chose from region_var, sex_var, year_var. Define multiple summarize variables by summarize_groups = c("region", "sex", "year")

summarize_icdcat
option to summarize over all ICD codes

output
Define the format of the output. Can be either "nested" for nested dataframe with fubreak_var and xbreak_var in separate sub_tables (purrr). Or "wide" for wide format where fubreak_var and xbreak_var are appended as columns. Or "long" for long format where sir_df is not reshaped, but just summarized (ybreak_var, xbreak_var and fubreak_var remain in rows). Default is "long".

output_information
option to define information to be presented in final output table. Default is "full" information, i.e. all variables from from sir_df. "reduced" is observed, expected, sir, sir_ci / sir_lci+sir_uci, pyar, n_base. "minimal" is observed, expected, sir, sir_ci. Default is "full".

add_total_row
option to add a row of totals. Can be either "no" for not adding such a row or "start" or "end" for adding it at the first or last row. Default is "no".

add_total_fu
option to add totals for follow-up time. Can be either "no" for not adding such a column or "start" or "end" for adding it at the first or last column. Default is "no".

collapse_ci
If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.

shorten_total_cols
Shorten text in all results columns that start with "Total". Default == FALSE.

fubreak_var_name
Name of variable with futime stratification. Default is "fu_time".

ybreak_var_name
Name of variable with futime stratification. Default is "yvar_name".

xbreak_var_name
Name of variable with futime stratification. Default is "xvar_name".

alpha
significance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.