Calculate crude incidence rates and cross-tabulate results by break variables; cumulative FU-times as are used as xbreak_var

Calculate crude incidence rates and cross-tabulate results by break variables; cumulative FU-times as are used as xbreak_var

ir_crosstab_byfutime(
  df,
  dattype = "zfkd",
  count_var,
  futime_breaks = c(0, 0.5, 1, 5, 10, Inf),
  ybreak_vars,
  collapse_ci = FALSE,
  add_total = "no",
  futime_var = NULL,
  alpha = 0.05
)

Arguments

df
dataframe in wide format

dattype
can be "zfkd" or "seer" or empty. Will set default variable names from dataset.

count_var
variable to be counted as observed case. Should be 1 for case to be counted.

futime_breaks
vector that indicates split points for follow-up time groups (in years) that will be used as xbreak_var. Default is c(0, .5, 1, 5, 10, Inf) that will result in 5 groups (up to 6 months, 6-12 months, 1-5 years, 5-10 years, 10+ years).

ybreak_vars
variables from df by which rates should be stratified in rows of result df. Multiple variables will result in appended rows in result df. y_break_vars is required.

collapse_ci
If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.

add_total
option to add a row of totals. Can be either "no" for not adding such a row or "top" or "bottom" for adding it at the first or last row. Default is "no".

futime_var
variable in df that contains follow-up time per person (in years). Default is set if dattype is given.

alpha
significance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.

Value

df

Examples

#load sample data
data("us_second_cancer")

#prep step - make wide data as this is the required format
usdata_wide <- us_second_cancer %>%
                    msSPChelpR::reshape_wide_tidyr(case_id_var = "fake_id", 
                    time_id_var = "SEQ_NUM", timevar_max = 10)
                    
#prep step - calculate p_spc variable
usdata_wide <- usdata_wide %>%
                 dplyr::mutate(p_spc = dplyr::case_when(is.na(t_site_icd.2)   ~ "No SPC",
                                                       !is.na(t_site_icd.2)   ~ "SPC developed",
                                                       TRUE ~ NA_character_)) %>%
                 dplyr::mutate(count_spc = dplyr::case_when(is.na(t_site_icd.2)   ~ 1,
                                                              TRUE ~ 0))
                                                              
#prep step - create patient status variable
usdata_wide <- usdata_wide %>%
                  msSPChelpR::pat_status(., fu_end = "2017-12-31", dattype = "seer",
                                         status_var = "p_status", life_var = "p_alive.1",
                                         birthdat_var = "datebirth.1", lifedat_var = "datedeath.1")

#> # A tibble: 11 x 3
#>    p_alive.1 p_status     n
#>    <chr>        <dbl> <int>
#>  1 Alive            1 15884
#>  2 Alive            2 17609
#>  3 Alive           97    19
#>  4 Alive           98  2488
#>  5 Dead             1  2544
#>  6 Dead             2  2147
#>  7 Dead             3 18358
#>  8 Dead             4  8822
#>  9 Dead            97     2
#> 10 Dead            98   182
#> 11 Dead            NA     5
#> # A tibble: 7 x 2
#>   p_status     n
#>      <dbl> <int>
#> 1        1 18428
#> 2        2 19756
#> 3        3 18358
#> 4        4  8822
#> 5       97    21
#> 6       98  2670
#> 7       NA     5

 
#now we can run the function
usdata_wide <- usdata_wide %>%
                 msSPChelpR::calc_futime(., 
                        futime_var_new = "p_futimeyrs", 
                        fu_end = "2017-12-31",
                        dattype = "seer", 
                        time_unit = "years",
                        status_var = "p_status",
                        lifedat_var = "datedeath.1", 
                        fcdat_var = "t_datediag.1", 
                        spcdat_var = "t_datediag.2")

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> # A tibble: 7 x 5
#>   p_status mean_futime min_futime max_futime median_futime
#>      <dbl>       <dbl>      <dbl>      <dbl>         <dbl>
#> 1        1        9.23     0.0438       27.0          8.04
#> 2        2        8.94     0            26.9          7.83
#> 3        3        8.60     0            25.8          7.55
#> 4        4        6.46     0            26.5          5.25
#> 5       97      NaN      Inf          -Inf           NA   
#> 6       98      NaN      Inf          -Inf           NA   
#> 7       NA      NaN      Inf          -Inf           NA   

                    
#for example, you can calculate incidence and summarize by sex and registry
msSPChelpR::ir_crosstab_byfutime(usdata_wide,
      dattype = "seer",
      count_var = "count_spc",
      futime_breaks = c(0, .5, 1, 5, 10, Inf),
      ybreak_vars = c("sex.1", "registry.1"),
      collapse_ci = FALSE,
      add_total = "no",
      futime_var = "p_futimeyrs",
      alpha = 0.05)

#> Warning: The variable for follow-up time has: 2762 missings. These will be omitted when creating the crosstabs.

#>    yvar_name                               yvar_label to 6 months_n_base
#> 1      sex.1                                   Female              32418
#> 2      sex.1                                     Male              32880
#> 3 registry.1 SEER Reg 01 - San Francisco-Oakland SMSA              21812
#> 4 registry.1                SEER Reg 02 - Connecticut              14513
#> 5 registry.1     SEER Reg 20 - Detroit (Metropolitan)              21836
#> 6 registry.1                     SEER Reg 21 - Hawaii               7137
#>   to 6 months_observed to 6 months_pyar to 6 months_abs_ir
#> 1                  997            15712            6345.66
#> 2                 1051            15926            6599.17
#> 3                  690            10563            6532.01
#> 4                  454             7034            6454.12
#> 5                  701            10573            6630.05
#> 6                  203             3467            5855.10
#>   to 6 months_abs_ir_lci to 6 months_abs_ir_uci 6-12 months_n_base
#> 1                5957.83                6752.11              30779
#> 2                6206.18                7010.51              31176
#> 3                6053.65                7038.11              20684
#> 4                5874.01                7076.03              13770
#> 5                6148.26                7139.55              20706
#> 6                5077.31                6718.33               6795
#>   6-12 months_observed 6-12 months_pyar 6-12 months_abs_ir
#> 1                  595            15075            3946.97
#> 2                  614            15272            4020.37
#> 3                  388            10137            3827.75
#> 4                  273             6748            4045.69
#> 5                  406            10142            4003.19
#> 6                  142             3321            4276.13
#>   6-12 months_abs_ir_lci 6-12 months_abs_ir_uci 1-5 years_n_base
#> 1                3636.16                4277.25            29513
#> 2                3708.61                4351.33            29902
#> 3                3456.31                4228.23            19858
#> 4                3579.96                4555.17            13212
#> 5                3623.22                4412.18            19856
#> 6                3601.75                5040.11             6489
#>   1-5 years_observed 1-5 years_pyar 1-5 years_abs_ir 1-5 years_abs_ir_lci
#> 1               4328         100154          4321.34              4193.54
#> 2               4489         101038          4442.90              4313.87
#> 3               2878          67430          4268.12              4113.59
#> 4               2022          44612          4532.41              4336.99
#> 5               2988          67249          4443.22              4285.31
#> 6                929          21901          4241.84              3973.42
#>   1-5 years_abs_ir_uci 5-10 years_n_base 5-10 years_observed 5-10 years_pyar
#> 1              4452.04             20830                4570           80883
#> 2              4574.81             20960                4611           81640
#> 3              4426.96             14064                3126           54439
#> 4              4734.37              9203                1999           35909
#> 5              4605.45             13963                3039           54500
#> 6              4523.62              4560                1017           17674
#>   5-10 years_abs_ir 5-10 years_abs_ir_lci 5-10 years_abs_ir_uci
#> 1           5650.12               5487.48               5816.35
#> 2           5647.96               5486.10               5813.38
#> 3           5742.16               5542.61               5947.06
#> 4           5566.81               5325.42               5816.31
#> 5           5576.11               5379.60               5777.96
#> 6           5754.10               5405.85               6118.90
#>   10+ years_n_base 10+ years_observed 10+ years_pyar 10+ years_abs_ir
#> 1            11998               6417          67402          9520.48
#> 2            12139               6278          68054          9225.09
#> 3             8088               4238          45328          9349.59
#> 4             5360               2863          30208          9477.63
#> 5             8101               4256          45561          9341.26
#> 6             2588               1338          14358          9318.79
#>   10+ years_abs_ir_lci 10+ years_abs_ir_uci
#> 1              9288.95              9756.33
#> 2              8998.29              9456.16
#> 3              9070.20              9635.41
#> 4              9133.61              9831.29
#> 5              9062.70              9626.20
#> 6              8826.09              9831.82