Calculate crude incidence rates and cross-tabulate results by break variables; cumulative FU-times as are used as xbreak_var

Calculate crude incidence rates and cross-tabulate results by break variables; cumulative FU-times as are used as xbreak_var

ir_crosstab_byfutime(
  df,
  dattype = "zfkd",
  count_var,
  futime_breaks = c(0, 0.5, 1, 5, 10, Inf),
  ybreak_vars,
  collapse_ci = FALSE,
  add_total = "no",
  futime_var = NULL,
  alpha = 0.05
)

Arguments

df
dataframe in wide format

dattype
can be "zfkd" or "seer" or empty. Will set default variable names from dataset.

count_var
variable to be counted as observed case. Should be 1 for case to be counted.

futime_breaks
vector that indicates split points for follow-up time groups (in years) that will be used as xbreak_var. Default is c(0, .5, 1, 5, 10, Inf) that will result in 5 groups (up to 6 months, 6-12 months, 1-5 years, 5-10 years, 10+ years).

ybreak_vars
variables from df by which rates should be stratified in rows of result df. Multiple variables will result in appended rows in result df. y_break_vars is required.

collapse_ci
If TRUE upper and lower confidence interval will be collapsed into one column separated by "-". Default is FALSE.

add_total
option to add a row of totals. Can be either "no" for not adding such a row or "top" or "bottom" for adding it at the first or last row. Default is "no".

futime_var
variable in df that contains follow-up time per person (in years). Default is set if dattype is given.

alpha
significance level for confidence interval calculations. Default is alpha = 0.05 which will give 95 percent confidence intervals.

Value

df

Examples

#load sample data
data("us_second_cancer")

#prep step - make wide data as this is the required format
usdata_wide <- us_second_cancer %>%
                    #only use sample
                    dplyr::filter(as.numeric(fake_id) < 200000) %>%
                    msSPChelpR::reshape_wide_tidyr(case_id_var = "fake_id", 
                    time_id_var = "SEQ_NUM", timevar_max = 2)

#> Long dataset had too many cases per patient. Wide dataset is limited to  2  cases per id as defined in timevar_max option.

                    
#prep step - calculate p_spc variable
usdata_wide <- usdata_wide %>%
                 dplyr::mutate(p_spc = dplyr::case_when(is.na(t_site_icd.2)   ~ "No SPC",
                                                       !is.na(t_site_icd.2)   ~ "SPC developed",
                                                       TRUE ~ NA_character_)) %>%
                 dplyr::mutate(count_spc = dplyr::case_when(is.na(t_site_icd.2)   ~ 1,
                                                              TRUE ~ 0))
                                                              
#prep step - create patient status variable
usdata_wide <- usdata_wide %>%
                  msSPChelpR::pat_status(., fu_end = "2017-12-31", dattype = "seer",
                                         status_var = "p_status", life_var = "p_alive.1",
                                         birthdat_var = "datebirth.1", lifedat_var = "datedeath.1")

#> # A tibble: 8 × 3
#>   p_alive.1 p_status     n
#>   <chr>        <dbl> <int>
#> 1 Alive            1  1728
#> 2 Alive            2  1896
#> 3 Alive           98   275
#> 4 Dead             1   275
#> 5 Dead             2   231
#> 6 Dead             3  2052
#> 7 Dead             4  1002
#> 8 Dead            98    27
#> # A tibble: 5 × 2
#>   p_status     n
#>      <dbl> <int>
#> 1        1  2003
#> 2        2  2127
#> 3        3  2052
#> 4        4  1002
#> 5       98   302

 
#now we can run the function
usdata_wide <- usdata_wide %>%
                 msSPChelpR::calc_futime(., 
                        futime_var_new = "p_futimeyrs", 
                        fu_end = "2017-12-31",
                        dattype = "seer", 
                        time_unit = "years",
                        status_var = "p_status",
                        lifedat_var = "datedeath.1", 
                        fcdat_var = "t_datediag.1", 
                        spcdat_var = "t_datediag.2")

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> # A tibble: 5 × 5
#>   p_status mean_futime min_futime max_futime median_futime
#>      <dbl>       <dbl>      <dbl>      <dbl>         <dbl>
#> 1        1        9.22     0.0438       27.0          8.13
#> 2        2        8.87     0            26.8          7.75
#> 3        3        8.39     0            25.5          7.38
#> 4        4        6.38     0            23.3          5.24
#> 5       98      NaN      Inf          -Inf           NA   

                    
#for example, you can calculate incidence and summarize by sex and registry
msSPChelpR::ir_crosstab_byfutime(usdata_wide,
      dattype = "seer",
      count_var = "count_spc",
      futime_breaks = c(0, .5, 1, 5, 10, Inf),
      ybreak_vars = c("sex.1", "registry.1"),
      collapse_ci = FALSE,
      add_total = "no",
      futime_var = "p_futimeyrs",
      alpha = 0.05)

#> Warning: The variable for follow-up time has: 309 missings. These will be omitted when creating the crosstabs.

#>    yvar_name                               yvar_label to 6 months_n_base
#> 1      sex.1                                   Female               3609
#> 2      sex.1                                     Male               3568
#> 3 registry.1 SEER Reg 01 - San Francisco-Oakland SMSA               2427
#> 4 registry.1                SEER Reg 02 - Connecticut               1573
#> 5 registry.1     SEER Reg 20 - Detroit (Metropolitan)               2398
#> 6 registry.1                     SEER Reg 21 - Hawaii                779
#>   to 6 months_observed to 6 months_pyar to 6 months_abs_ir
#> 1                  106             1749            6060.98
#> 2                  116             1728            6711.18
#> 3                   66             1180            5594.77
#> 4                   62              757            8185.10
#> 5                   68             1165            5836.42
#> 6                   26              375            6931.39
#>   to 6 months_abs_ir_lci to 6 months_abs_ir_uci 6-12 months_n_base
#> 1                4962.24                7330.58               3428
#> 2                5545.58                8049.42               3376
#> 3                4326.99                7117.92               2315
#> 4                6275.47               10492.92               1476
#> 5                4532.21                7399.05               2283
#> 6                4527.81               10156.09                730
#>   6-12 months_observed 6-12 months_pyar 6-12 months_abs_ir
#> 1                   69             1675            4120.30
#> 2                   71             1652            4298.88
#> 3                   36             1135            3172.12
#> 4                   39              718            5428.77
#> 5                   52             1117            4655.93
#> 6                   13              356            3650.73
#>   6-12 months_abs_ir_lci 6-12 months_abs_ir_uci 1-5 years_n_base
#> 1                3205.84                5214.50             3274
#> 2                3357.46                5422.45             3234
#> 3                2221.71                4391.55             2227
#> 4                3860.39                7421.31             1396
#> 5                3477.27                6105.64             2188
#> 6                1943.86                6242.86              697
#>   1-5 years_observed 1-5 years_pyar 1-5 years_abs_ir 1-5 years_abs_ir_lci
#> 1                484          11057          4377.15              3995.83
#> 2                501          10861          4613.04              4217.89
#> 3                323           7588          4256.90              3805.27
#> 4                228           4666          4886.32              4272.60
#> 5                327           7383          4428.80              3961.73
#> 6                107           2281          4691.55              3844.84
#>   1-5 years_abs_ir_uci 5-10 years_n_base 5-10 years_observed 5-10 years_pyar
#> 1              4785.04              2286                 536            8848
#> 2              5035.24              2260                 514            8781
#> 3              4747.39              1585                 392            6185
#> 4              5563.44               951                 200            3651
#> 5              4935.80              1533                 351            5980
#> 6              5669.26               477                 107            1813
#>   5-10 years_abs_ir 5-10 years_abs_ir_lci 5-10 years_abs_ir_uci
#> 1           6057.70               5555.66               6592.93
#> 2           5853.54               5358.37               6382.16
#> 3           6338.12               5726.15               6997.70
#> 4           5478.13               4745.18               6292.23
#> 5           5869.25               5271.23               6516.54
#> 6           5901.01               4836.02               7130.77
#>   10+ years_n_base 10+ years_observed 10+ years_pyar 10+ years_abs_ir
#> 1             1294                675           7277          9276.12
#> 2             1292                687           7261          9461.50
#> 3              891                458           5034          9098.18
#> 4              539                302           3060          9870.53
#> 5              891                474           4912          9649.92
#> 6              265                128           1532          8353.96
#>   10+ years_abs_ir_lci 10+ years_abs_ir_uci
#> 1              8589.44             10003.08
#> 2              8767.13             10196.25
#> 3              8283.91              9970.84
#> 4              8788.57             11048.91
#> 5              8800.63             10559.03
#> 6              6969.51              9932.89